backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis[hiredis]>=5.0.0
celery[redis]>=5.4.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
tenacity>=8.3.0
python-multipart>=0.0.9
slowapi>=0.1.9
